arrumado3

+++ src/main/java/br/com/db1/rest/CalculadoraRest.java
@@ -15,8 +15,8 @@ public class CalculadoraRest {
 	
 	@GET
 	@Path("dividir/{param}/{param2}")
-	public Response calcular(@PathParam("param") Double valor1, @PathParam("param2") Double valor2) {
-		Double result = dao.dividir(valor1, valor2);
+	public Response calcular(@PathParam("param") Integer valor1, @PathParam("param2") Integer valor2) {
+		Integer result = dao.dividir(valor1, valor2);
 		return Response.status(200).entity(result).build();
 	}
 }
